PluggableDataIO: Move importable factory to toolkit
