+++ caikit/core/toolkit/factory.py
@@ -19,6 +19,7 @@ base class for classes that can be constructed via caikit config
 # Standard
 from typing import Optional, Type
 import abc
+import importlib
 
 # First Party
 import aconfig
@@ -105,3 +106,56 @@ class Factory:
         )
         instance_name = instance_name or inst_cls.name
         return inst_cls(inst_cfg, instance_name)
+
+
+class ImportableFactory(Factory):
+    """An ImportableFactory extends the base Factory to allow the construction
+    to specify an "import_class" field that will be used to import and register
+    the implementation class before attempting to initialize it.
+    """
+
+    IMPORT_CLASS_KEY = "import_class"
+
+    def construct(
+        self,
+        instance_config: dict,
+        instance_name: Optional[str] = None,
+    ):
+        # Look for an import_class and import and register it if found
+        import_class_val = instance_config.get(self.__class__.IMPORT_CLASS_KEY)
+        if import_class_val:
+            error.type_check(
+                "<COR85108801E>",
+                str,
+                **{self.__class__.IMPORT_CLASS_KEY: import_class_val},
+            )
+            module_name, class_name = import_class_val.rsplit(".", 1)
+            try:
+                imported_module = importlib.import_module(module_name)
+            except ImportError:
+                error(
+                    "<COR46837141E>",
+                    ValueError(
+                        "Invalid {}: Module cannot be imported [{}]".format(
+                            self.__class__.IMPORT_CLASS_KEY,
+                            module_name,
+                        )
+                    ),
+                )
+            try:
+                imported_class = getattr(imported_module, class_name)
+            except AttributeError:
+                error(
+                    "<COR46837142E>",
+                    ValueError(
+                        "Invalid {}: No such class [{}] on module [{}]".format(
+                            self.__class__.IMPORT_CLASS_KEY,
+                            class_name,
+                            module_name,
+                        )
+                    ),
+                )
+            error.subclass_check("<COR52306423E>", imported_class, FactoryConstructible)
+
+            self.register(imported_class)
+        return super().construct(instance_config, instance_name)
